requirements.txt
click==8.1.3
dnspython==2.2.1
dominate==2.6.0
email-validator==1.2.1
Flask==2.1.2
Flask-Bootstrap4==4.0.2
Flask-Cors==3.0.10
flask-mongoengine==1.0.0
Flask-WTF==1.0.1
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mongoengine==0.24.1
pymongo==4.1.1
python-dotenv==0.20.0
six==1.16.0
visitor==0.1.3
Werkzeug==2.1.2
WTForms==3.0.1
